# ticket_management_system/swagger_specs/swagger.yml
swagger: '2.0'
info:
  title: Flight Management System API
  description: RESTful API for managing flights and bookings with scoped user-ID tokens
  version: 1.0.0
basePath: /api
host: localhost:8080
schemes:
- http
- https
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      JWT Authorization header using the Bearer scheme.
      Example: "Authorization: Bearer {token}"
parameters:
  PageParam:
    name: page
    in: query
    type: integer
    default: 1
    description: Page number (1-indexed)
  PerPageParam:
    name: per_page
    in: query
    type: integer
    default: 10
    description: Number of items per page (max 100)
  BookingIdParam:
    name: booking_id
    in: path
    required: true
    type: string
    format: uuid
    description: Booking identifier
...
definitions:
...
  ValidationError:
    type: object
    properties:
      error:
        type: string
        example: Bad Request
      message:
        type: string
        example: Validation failed
      errors:
        type: object
        additionalProperties: true
        example:
          email:
          - Not a valid email address
          password:
          - Password must be at least 6 characters
    required:
    - error
    - message
  AuthError:
    type: object
    properties:
      error:
        type: string
        example: Unauthorized
      message:
        type: string
        example: Invalid or missing authentication token
    required:
    - error
    - message
  NotFoundError:
    type: object
    properties:
      error:
        type: string
        example: Not Found
      message:
        type: string
        example: Resource not found
    required:
    - error
    - message
  ConflictError:
    type: object
    properties:
      error:
        type: string
        example: Conflict
      message:
        type: string
        example: Resource already exists or operation cannot be performed
    required:
    - error
    - message
  ForbiddenError:
    type: object
    properties:
      error:
        type: string
        example: Forbidden
      message:
        type: string
        example: You do not have permission to access this resource
    required:
    - error
    - message
...
  Ticket:
    type: object
    properties:
      id:
        type: string
        format: uuid
      passenger_name:
        type: string
        description: Legacy full name assembled from passenger_fname and passenger_lname.
      passenger_fname:
        type: string
      passenger_lname:
        type: string
      passenger_passport_num:
        type: string
      seat_num:
        type: string
      seat_class:
        type: string
        enum:
        - economy
        - business
        - first
      price:
        type: string
      created_at:
        type: string
        format: date-time
  Booking:
    type: object
    properties:
      id:
        type: string
        format: uuid
      user_id:
        type: string
        format: uuid
        x-nullable: true
        description: Owner user identifier. Null for anonymous bookings.
      flight_id:
        type: string
        format: uuid
      total_price:
        type: string
      booking_status:
        type: string
        enum:
        - booked
        - paid
        - cancelled
        - refunded
      ticket_count:
        type: integer
        description: Number of tickets (only in list views)
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
      tickets:
        type: array
        description: Full ticket details (only in detail views)
        items:
          $ref: '#/definitions/Ticket'
  BookingResponse:
    type: object
    properties:
      message:
        type: string
      booking:
        $ref: '#/definitions/Booking'
...
  Pagination:
    type: object
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total_pages:
        type: integer
      total_items:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      next_page:
        type: integer
        nullable: true
      prev_page:
        type: integer
        nullable: true
...
  BookingListResponse:
    type: object
    properties:
      bookings:
        type: array
        items:
          $ref: '#/definitions/Booking'
      pagination:
        $ref: '#/definitions/Pagination'
...
  SeatAvailabilityResponse:
    type: object
    properties:
      flight_id:
        type: string
        format: uuid
      seat_num:
        type: string
      available:
        type: boolean
paths:
...
  /bookings:
    post:
      summary: Create a new booking
      description: Create a flight booking with one or more passengers. Authentication is not required. The first passenger email is required for booking ownership and receipt delivery; it is used to assign the booking to an existing user or create a new user.
      operationId: createBooking
      tags:
      - Bookings
      parameters:
      - in: body
        name: body
        required: true
        schema:
          type: object
          required:
          - flight_id
          - passengers
          properties:
            user_id:
              type: string
              format: uuid
              x-nullable: true
              description: Optional user identifier for clients that want to associate the booking with an existing user.
            flight_id:
              type: string
              format: uuid
            passengers:
              type: array
              minItems: 1
              description: First passenger must include email. Additional passenger emails are optional.
              items:
                type: object
                required:
                - passenger_fname
                - passenger_lname
                - passenger_passport_num
                - seat_num
                properties:
                  passenger_name:
                    type: string
                    description: Legacy alternative to passenger_fname/passenger_lname.
                  passenger_fname:
                    type: string
                  passenger_lname:
                    type: string
                  email:
                    type: string
                    format: email
                    description: Required on the first passenger for booking ownership and receipt delivery; optional for additional passengers.
                  passenger_passport_num:
                    type: string
                  seat_num:
                    type: string
                  seat_class:
                    type: string
                    enum:
                    - economy
                    - business
                    - first
      responses:
        201:
          description: Booking created successfully
          schema:
            $ref: '#/definitions/BookingResponse'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
        404:
          description: Flight or user not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
        409:
          description: Conflict (e.g. seat already taken, or flight not in a bookable state)
          schema:
            $ref: '#/definitions/ConflictError'
          examples:
            application/json: *conflict_error_example
    get:
      summary: List bookings
      description: List bookings with pagination. Regular token users receive only their own bookings; admins receive all bookings.
      operationId: listBookings
      tags:
      - Bookings
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/PageParam'
      - $ref: '#/parameters/PerPageParam'
      responses:
        200:
          description: Paginated list of bookings
          schema:
            $ref: '#/definitions/BookingListResponse'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - missing bookings:read permission
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: *forbidden_error_example
  /bookings/{booking_id}:
    get:
      summary: Get booking details
      description: Retrieve details of a specific booking. Regular token users can access their own bookings; admins can access any booking.
      operationId: getBooking
      tags:
      - Bookings
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/BookingIdParam'
      responses:
        200:
          description: Booking details
          schema:
            type: object
            properties:
              booking:
                $ref: '#/definitions/Booking'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - booking does not belong to the token user
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: *forbidden_error_example
        404:
          description: Booking not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
    put:
      summary: Update booking status
      description: Update the status of an existing booking. Regular token users can update their own bookings; admins can update any booking.
      operationId: updateBooking
      tags:
      - Bookings
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/BookingIdParam'
      - in: body
        name: body
        required: true
        schema:
          type: object
          required:
          - booking_status
          properties:
            booking_status:
              type: string
              enum:
              - booked
              - paid
              - cancelled
              - refunded
      responses:
        200:
          description: Booking updated successfully
          schema:
            $ref: '#/definitions/BookingResponse'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - booking does not belong to the token user
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: *forbidden_error_example
        404:
          description: Booking not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
        409:
          description: Conflict (e.g. booking already cancelled or refunded)
          schema:
            $ref: '#/definitions/ConflictError'
          examples:
            application/json: *conflict_error_example
    delete:
      summary: Cancel booking
      description: Cancel an existing booking by setting its status to cancelled. Regular token users can cancel their own bookings; admins can cancel any booking.
      operationId: cancelBooking
      tags:
      - Bookings
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/BookingIdParam'
      responses:
        200:
          description: Booking cancelled successfully
          schema:
            $ref: '#/definitions/BookingResponse'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - booking does not belong to the token user
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: *forbidden_error_example
        404:
          description: Booking not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
        409:
          description: Conflict (e.g. booking already cancelled or cannot cancel refunded booking)
          schema:
            $ref: '#/definitions/ConflictError'
          examples:
            application/json: *conflict_error_example
  /bookings/availability:
    get:
      summary: Check seat availability
      description: Check if a specific seat is available on a given flight. Authentication is not required.
      operationId: checkAvailability
      tags:
      - Bookings
      parameters:
      - name: flight_id
        in: query
        required: true
        type: string
        format: uuid
      - name: seat_num
        in: query
        required: true
        type: string
      responses:
        200:
          description: Seat availability result
          schema:
            $ref: '#/definitions/SeatAvailabilityResponse'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
...